+++ R/gg-bin2d.R
@@ -32,8 +32,8 @@ gg_bin2d <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = FALSE,
+  
+  border = FALSE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -200,7 +200,7 @@ gg_bin2d <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},