+++ R/gg-col.R
@@ -42,8 +42,8 @@ gg_col <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = TRUE,
+  
+  border = TRUE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -210,7 +210,7 @@ gg_col <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},