+++ R/gg-contour-filled.R
@@ -27,8 +27,8 @@ gg_contour_filled <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = FALSE,
+  
+  border = FALSE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -195,7 +195,7 @@ gg_contour_filled <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},