+++ R/gg-contour.R
@@ -24,8 +24,8 @@ gg_contour <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = FALSE,
+  
+  border = FALSE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -192,7 +192,7 @@ gg_contour <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},