+++ R/gg-crossbar.R
@@ -62,8 +62,8 @@ gg_crossbar <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = TRUE,
+  
+  border = TRUE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -230,7 +230,7 @@ gg_crossbar <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},