+++ R/gg-density2d.R
@@ -30,8 +30,8 @@ gg_density2d <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = FALSE,
+  
+  border = FALSE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -198,7 +198,7 @@ gg_density2d <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},