+++ R/gg-jitter.R
@@ -37,8 +37,8 @@ gg_jitter <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = TRUE,
+  
+  border = TRUE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -205,7 +205,7 @@ gg_jitter <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},