+++ R/gg-point.R
@@ -73,8 +73,8 @@ gg_point <- function(
   before = NULL,
   with = NULL,
   
-  refine = NULL,
-  border = TRUE,
+  
+  border = TRUE, theme = NULL, refine = NULL,
   x = NULL,
   xmin = NULL,
   xmax = NULL,
@@ -241,7 +241,7 @@ gg_point <- function(
     before = before,
     with = with,
     
-    refine = refine,
+    theme = theme, refine = refine,
     border = border,
     x = {{ x }},
     xmin = {{ xmin }},